fix: 32 bit errors and add siplus_type_new_s
siplus_type_new_s was required due to some wasm errors in the C# bindings.
Only a maximum of 6 parameters were able to be used, so we needed a shorter
call signature.

Side note: Maybe we should add _s variants for all create/new functions in the C API?

+++ siplus/stdlib/text.cpp
@@ -423,7 +423,9 @@ string_comparator::invoke(
     } else if(strl > strr) {
         return make_data<types::IntegerType>(1L);
     } else {
-        return make_data<types::IntegerType>(0L);
+        //static_cast needed because 0 (specifically 0) is convertible to nullptr, 
+        //and therefore a pointer type and causes ambiguity in make_data
+        return make_data<types::IntegerType>(static_cast<int64_t>(0));
     }
 }
